Fixed long dash in artist list outputfield/outputfield#41
