+++ components/artists/artistRow.component.tsx
@@ -2,6 +2,7 @@ import React from 'react'
 import Link from 'next/link'
 import { ArtistWithInviterAndUserAndLinks } from '../../pages/api/artists/[name]'
 import ProfilePicture from './profilePicture'
+import DashedDivider from '../dashedDivider'
 
 export interface ArtistRowProps {
   artist: ArtistWithInviterAndUserAndLinks;
@@ -16,10 +17,9 @@ export const ArtistRow = ({ artist, type }: ArtistRowProps) => {
     <div
       data-testid='artistRow'
       className={`
-        ${type === 'detail' ? 'border-long-dashed-y' : ''} 
-        ${type === 'list' ? 'border-long-dashed-t' : ''} 
-        w-full flex flex-col relative pt-[11px] pl-[13px] pb-[8px] pr-[11px]`}>
-      <div className="grow relative flex justify-center self-start items-center">
+        w-full flex flex-col relative `}>
+      <DashedDivider />
+      <div className="grow relative flex justify-center self-start items-center pt-[11px] pl-[13px] pb-[8px] pr-[11px]">
         <div className="mx-[10px] my-[24px]">
           <ProfilePicture />
           {type == 'detail' && <div className="text-center font-serif mt-1">{artist?.pronouns}</div>}
@@ -36,13 +36,14 @@ export const ArtistRow = ({ artist, type }: ArtistRowProps) => {
       <span className="text-right text-lg uppercase absolute right-[13px] bottom-[8px]">
         {artist?.mediums}
       </span>
+      {type === 'detail' && <DashedDivider />}
     </div>
   )
   if (type === 'list') {
     return (
       <Link
         href={uri}
-        className="last-of-type:border-long-dashed-y"
+        className='cursor-pointer'
         legacyBehavior>
         {row}
       </Link>
